+++ server/package.json
@@ -27,10 +27,12 @@
     "express": "~4.16.1",
     "express-session": "^1.17.1",
     "http-errors": "~1.6.3",
+    "jsonwebtoken": "^8.5.1",
     "morgan": "^1.10.0",
     "mysql2": "^2.1.0",
     "passport": "^0.4.1",
     "passport-github2": "^0.1.12",
+    "passport-jwt": "^4.0.0",
     "path": "^0.12.7",
     "sequelize": "^6.3.4",
     "session-file-store": "^1.4.0",
@@ -46,11 +48,13 @@
     "@types/elasticsearch": "^5.0.36",
     "@types/express": "^4.17.7",
     "@types/express-session": "^1.17.0",
+    "@types/jsonwebtoken": "^8.5.0",
     "@types/mocha": "^8.0.1",
     "@types/morgan": "^1.9.1",
     "@types/node": "^14.0.26",
     "@types/passport": "^1.0.4",
     "@types/passport-github2": "^1.2.4",
+    "@types/passport-jwt": "^3.0.3",
     "@types/session-file-store": "^1.2.1",
     "@types/supertest": "^2.0.10",
     "@types/tunnel-ssh": "^4.1.0",
